Test file in which I tried stuff
I tried to implement something in which it would keep track of what had already been tried with a 3d array but it keeps looping for some reason. It also at some point stopped recognizing the thread sleep command which I think is weird. I think this can work if implemented correctly but for now it just keeps looping.